finish Bronze & Silver Challenge
population: 5922, number of stoplights: 4

Dracula is terrorizing a town!
There are 1 vampires in this town.

Dracula is terrorizing a town!
There are 2 vampires in this town.

Dracula is terrorizing a town!
There are 3 vampires in this town.

population: 5919, number of stoplights: 4

Zombie is terrorizing a town!
population: 4922, number of
stoplights: 4